*KVM: SEV: Add KVM_SNP_INIT command
TODO: move to using common KVM_SEV_INIT2 interface. The flag discovery
mechanism also needs to be changed since it provides no mechanism for
discovery outside of trying to trigger the error path by guessing at
unimplemented flags. May went to leverage something similar to the
KVM_SEV_INIT2 VMSA feature machinery for this as well which involves a
KVM device ioctl to return the feature bitmap.

The KVM_SNP_INIT command is used by the hypervisor to initialize the
SEV-SNP platform context. In a typical workflow, this command should be
the first command issued. When creating SEV-SNP guest, the VMM must use
this command instead of the KVM_SEV_INIT or KVM_SEV_ES_INIT.

The flags value must be zero, it will be extended in future SNP support
to communicate the optional features (such as restricted INT injection
etc).

Co-developed-by: Pavan Kumar Paluri <papaluri@amd.com>
Signed-off-by: Pavan Kumar Paluri <papaluri@amd.com>
Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>
Signed-off-by: Ashish Kalra <ashish.kalra@amd.com>
Signed-off-by: Michael Roth <michael.roth@amd.com>